move shortest path functions to distance module
